Connect mutate function to websocket through socket-parser
This means that the display of the current planet does get updated from
the websocket connection, however this is not the "om-next way" but
an intermediate step before connecting them through a core.async channel.

The socket state and/or callback function probably shouldn't know
anything about the mutate query.