introduce set-remote-id-callback in App and pass through to children
this is what fixed the recursive updates - doing om/trasact! in the
context of a child leaves om unable to track the UI state changes when
a mutation affects another component on the sibling level

instead, the transact! to set the remote-id is called from the body of
the top-level App component, and is passed through to the child
components so that they can call it in their lifecycle methods

https://github.com/awkay/om/wiki/Om-Next-Overview#locating-calls-to-transact